+++ src/core/index.ts
@@ -29,6 +29,7 @@ export * from "./events/index.js";
 export * from "./lights/index.js";
 export * from "./materials/index.js";
 export * from "./textures/index.js";
+export * from "./text/index.js";
 export * from "./fsm/index.js";
 export * from "./threading/index.js";